requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pytesseract==0.3.10
Pillow==10.0.1
PyPDF2==3.0.1
PyMuPDF==1.23.5
python-docx==0.8.11
requests==2.31.0
